pyproject.toml
[tool.poetry]
name = "Feecc-Hub"
version = "0.1.0"
description = "Hub software for the Feecc QA system"
authors = ["arseniiarsenii <arseniivelichko2@gmail.com>"]
license = "Apache 2.0"

[tool.poetry.dependencies]
python = ">3.7.0,<3.7.1 || >3.7.1,<4.0"
requests = "^2.25.1"
Pillow = "^8.2.0"
PyYAML = "^5.4.1"
brother-ql = "^0.9.4"
qrcode = "^6.1"
python-barcode = "^0.13.1"
fastapi = "^0.66.0"
uvicorn = "^0.14.0"
dnspython = "^2.1.0"
substrate-interface = "^0.13.12"
loguru = "^0.5.3"
motor = "^2.5.1"
httpx = "^0.19.0"

[tool.poetry.dev-dependencies]
mypy = "^0.902"
black = "^21.6b0"
flake8 = "^3.9.2"
types-PyYAML = "^5.4.3"
types-requests = "^2.25.0"
pytest = "^6.2.4"
vulture = "^2.3"
pre-commit = "^2.15.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
exclude = "test*"
strict = true
ignore_missing_imports  = true
allow_untyped_decorators = true

[tool.black]
line-length = 120
